+++ plugin.video.dailymotion_com/addon.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<addon id="plugin.video.dailymotion_com" name="DailyMotion.com" version="2.4.0" provider-name="gujal">
+<addon id="plugin.video.dailymotion_com" name="DailyMotion.com" version="2.4.1" provider-name="gujal">
     <requires>
         <import addon="xbmc.python" version="2.24.0"/>
 		<import addon="script.module.future" version="0.0.1"/>
@@ -24,10 +24,9 @@
 		<disclaimer lang="he">ייתכן שחלק מהתוספות האלו אינן חוקיות בארץ המגורים שלך - אנא בדוק את החוקים המקומיים שלך לפני ההתקנה.</disclaimer>
 	    
         <news>
-[B]Version 2.4.0 - 08 Dec 2019[/B] 
- - Fix for playback due to api changes
- - Flake8 linting
- - Known issue: cannot download videos anymore
+[B]Version 2.4.1 - 17 Dec 2019[/B] 
+ - Fix Search (Thanks to johngf)
+ - Add Search keyword History (Thanks to johngf)
 		</news>
 		<website>https://www.dailymotion.com</website>
         <source>https://github.com/Gujal00/plugin.video.dailymotion_com</source>
